Add new data instruction generation and control flow instruction handling; implement templates for various instructions and enhance type validation

+++ src/main/java/dev/superice/gdcc/lir/LirInstruction.java
@@ -55,7 +55,7 @@ public interface LirInstruction {
         }
     }
 
-    record FloatOperand(float value) implements Operand {
+    record FloatOperand(double value) implements Operand {
         @Override
         public OperandKind operandKind() {
             return OperandKind.FLOAT;
